pyproject.toml
[project]
name = "deep-research-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.15",
    "autogen-agentchat>=0.7.4",
    "autogen-core>=0.7.4",
    "autogen-ext[azure]>=0.7.4",
    "ddgs>=9.6.0",
    "dotenv>=0.9.9",
    "duckduckgo-search>=8.1.1",
    "lxml-html-clean>=0.4.2",
    "newspaper4k>=0.9.3.1",
    "nltk>=3.9.1",
    "python-dotenv>=1.1.1",
]
